fix jsshell perf testing
