+++ tools/perf/benchmark.jsshell.js
@@ -1,10 +1,7 @@
-load('../../dist/jsshell/l10n.js');
+load('../../build/dist/jsshell/l20n.js');
 
-var propertiesParser = L20n.PropertiesParser;
+//var propertiesParser = L20n.PropertiesParser;
 var l20nParser = L20n.L20nParser;
-var env = {
-  __plural: L20n.getPluralRule('en-US')
-};
 
 var propCode = read('./example.properties');
 var l20nCode = read('./example.l20n');
@@ -26,7 +23,13 @@ var data = {
   "number": "NUMBER",
   "link1": "LINK1",
   "link2": "LINK2"
-}
+};
+
+var lang = {
+  code:'en-US',
+  src: 'app',
+  dir: 'ltr'
+};
 
 function micro(time) {
   // time is in milliseconds with decimals
@@ -36,8 +39,8 @@ function micro(time) {
 var times = {};
 times.start = dateNow();
 
-var ast = propertiesParser.parse(null, propCode);
-times.parseEnd = dateNow();
+//var ast = propertiesParser.parse(null, propCode);
+//times.parseEnd = dateNow();
 
 times.l20nParseStart = dateNow();
 
@@ -45,14 +48,14 @@ var ast = l20nParser.parse(null, l20nCode);
 times.l20nParseEnd = dateNow();
 
 times.createEntries = dateNow();
-L20n.extendEntries(env, ast);
+var entries = L20n.createEntriesFromAST(ast);
 times.createEntriesEnd = dateNow();
 
-var ids = Object.keys(env).filter(function(id){return id !== '__plural';});
+var ctx = new L20n.MockContext(entries);
 
 times.format = dateNow();
-for (var id in ids) {
-   L20n.Resolver.format(data, env[ids[id]]);
+for (var id in entries) {
+   L20n.format(ctx, lang, data, entries[id]);
 }
 times.formatEnd = dateNow();
 /*
@@ -68,7 +71,7 @@ for (var id in ids) {
 times.getEntityEnd = dateNow();
 */
 var results = {
-  parseProp: micro(times.parseEnd - times.start),
+  //parseProp: micro(times.parseEnd - times.start),
   parseL20n: micro(times.l20nParseEnd - times.l20nParseStart),
   createEntries: micro(times.createEntriesEnd - times.createEntries),
   format: micro(times.formatEnd - times.format),
